Reorganize vertex and fragment shader
This is in preparation for the next steps, it's just refactoring the
vertex shader to store all quads that need rendering in one global array
and using mod and div to know which one to draw. And also passing the
color from that quad data to the fragment shader.
Later on, this data will be passed to the vertex shader through a
storage buffer