Add drawQuad
This is a big commit, but basically what's changed is that now, there's
a drawQuad function, to which you pass quad data. Every time you call
this function, a new quad data element is added to an array, which
is then memcopied into the storage buffer.
What's left now is texture fill and the SDF functionality (rounded
corners, strokes, etc.)