Get blending to work and fix weird artifacts!!
Getting blending to work was pretty straightforward, the bigger
problem was the bug where weird little artifacts would show up all over
the screen when I tried rendering more than 1 texture.
In the end, all I needed to do was add the GL_EXT_nonuniform_qualifier
extension in the fragment shader glsl and instead of indexing directly
into the textureSamplers array, I call nonUniformEXT on the index and
use that (again in the fragment shader).
I asked about this in the "Graphics Programming" discord server when
trying to find the solution, here's the invite URL:
https://discord.gg/6mgNGk7