Finalize rounded corners implementation
The draw commands and structs have been changed a lot, as well as the
vertex and fragment shaders, but the core difference is that now
there's a DrawQuadOptions parameter that lets you specify the four
corner radii (and opacity).
Also rounded corners work on texture fills too now, not just color
fills.